Crate rusoto_transcribe

Operations and objects for transcribing speech to text.

If you're using the service, you're probably looking for TranscribeClient and Transcribe.

Structs

CreateVocabularyRequest

CreateVocabularyResponse

DeleteTranscriptionJobRequest

DeleteVocabularyRequest

GetTranscriptionJobRequest

GetTranscriptionJobResponse

GetVocabularyRequest

GetVocabularyResponse

ListTranscriptionJobsRequest

ListTranscriptionJobsResponse

ListVocabulariesRequest

ListVocabulariesResponse

Media
Describes the input media file in a transcription request.

Settings
Provides optional settings for the StartTranscriptionJob operation.

StartTranscriptionJobRequest

StartTranscriptionJobResponse

TranscribeClient
A client for the Amazon Transcribe Service API.

Transcript
Identifies the location of a transcription.

TranscriptionJob
Describes an asynchronous transcription job that was created with the StartTranscriptionJob operation.

TranscriptionJobSummary
Provides a summary of information about a transcription job. .

UpdateVocabularyRequest

UpdateVocabularyResponse

VocabularyInfo
Provides information about a custom vocabulary.

Enums

CreateVocabularyError
Errors returned by CreateVocabulary

DeleteTranscriptionJobError
Errors returned by DeleteTranscriptionJob

DeleteVocabularyError
Errors returned by DeleteVocabulary

GetTranscriptionJobError
Errors returned by GetTranscriptionJob

GetVocabularyError
Errors returned by GetVocabulary

ListTranscriptionJobsError
Errors returned by ListTranscriptionJobs

ListVocabulariesError
Errors returned by ListVocabularies

StartTranscriptionJobError
Errors returned by StartTranscriptionJob

UpdateVocabularyError
Errors returned by UpdateVocabulary

Traits

Transcribe
Trait representing the capabilities of the Amazon Transcribe Service API. Amazon Transcribe Service clients implement this trait.