Struct rusoto_transcribe::TranscribeClient

pub struct TranscribeClient { /* fields omitted */ }

A client for the Amazon Transcribe Service API.

Methods

impl TranscribeClient

pub fn new(region: Region) -> TranscribeClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> TranscribeClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,