Trait rusoto_transcribe::Transcribe

pub trait Transcribe {
    fn create_vocabulary(
        &self, 
        input: CreateVocabularyRequest
    ) -> RusotoFuture<CreateVocabularyResponse, CreateVocabularyError>;

    fn delete_transcription_job(
        &self, 
        input: DeleteTranscriptionJobRequest
    ) -> RusotoFuture<(), DeleteTranscriptionJobError>;

    fn delete_vocabulary(
        &self, 
        input: DeleteVocabularyRequest
    ) -> RusotoFuture<(), DeleteVocabularyError>;

    fn get_transcription_job(
        &self, 
        input: GetTranscriptionJobRequest
    ) -> RusotoFuture<GetTranscriptionJobResponse, GetTranscriptionJobError>;

    fn get_vocabulary(
        &self, 
        input: GetVocabularyRequest
    ) -> RusotoFuture<GetVocabularyResponse, GetVocabularyError>;

    fn list_transcription_jobs(
        &self, 
        input: ListTranscriptionJobsRequest
    ) -> RusotoFuture<ListTranscriptionJobsResponse, ListTranscriptionJobsError>;

    fn list_vocabularies(
        &self, 
        input: ListVocabulariesRequest
    ) -> RusotoFuture<ListVocabulariesResponse, ListVocabulariesError>;

    fn start_transcription_job(
        &self, 
        input: StartTranscriptionJobRequest
    ) -> RusotoFuture<StartTranscriptionJobResponse, StartTranscriptionJobError>;

    fn update_vocabulary(
        &self, 
        input: UpdateVocabularyRequest
    ) -> RusotoFuture<UpdateVocabularyResponse, UpdateVocabularyError>;
}

Trait representing the capabilities of the Amazon Transcribe Service API. Amazon Transcribe Service clients implement this trait.

Required methods

fn create_vocabulary(
&self,
input: CreateVocabularyRequest
) -> RusotoFuture<CreateVocabularyResponse, CreateVocabularyError>

Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an audio file.

fn delete_transcription_job(
&self,
input: DeleteTranscriptionJobRequest
) -> RusotoFuture<(), DeleteTranscriptionJobError>

Deletes a previously submitted transcription job along with any other generated results such as the transcription, models, and so on.

fn delete_vocabulary(
&self,
input: DeleteVocabularyRequest
) -> RusotoFuture<(), DeleteVocabularyError>

Deletes a vocabulary from Amazon Transcribe.

fn get_transcription_job(
&self,
input: GetTranscriptionJobRequest
) -> RusotoFuture<GetTranscriptionJobResponse, GetTranscriptionJobError>

Returns information about a transcription job. To see the status of the job, check the TranscriptionJobStatus field. If the status is COMPLETED, the job is finished and you can find the results at the location specified in the TranscriptionFileUri field.

fn get_vocabulary(
&self,
input: GetVocabularyRequest
) -> RusotoFuture<GetVocabularyResponse, GetVocabularyError>

Gets information about a vocabulary.

fn list_transcription_jobs(
&self,
input: ListTranscriptionJobsRequest
) -> RusotoFuture<ListTranscriptionJobsResponse, ListTranscriptionJobsError>

Lists transcription jobs with the specified status.

fn list_vocabularies(
&self,
input: ListVocabulariesRequest
) -> RusotoFuture<ListVocabulariesResponse, ListVocabulariesError>

Returns a list of vocabularies that match the specified criteria. If no criteria are specified, returns the entire list of vocabularies.

fn start_transcription_job(
&self,
input: StartTranscriptionJobRequest
) -> RusotoFuture<StartTranscriptionJobResponse, StartTranscriptionJobError>

Starts an asynchronous job to transcribe speech to text.

fn update_vocabulary(
&self,
input: UpdateVocabularyRequest
) -> RusotoFuture<UpdateVocabularyResponse, UpdateVocabularyError>

Updates an existing vocabulary with new values. The UpdateVocabulary operation overwrites all of the existing information with the values that you provide in the request.

Loading content...